ISSUE #2
Separate updating of ThingSpeak (and handling of esp8266) from main
file (ino), to make code more clear.

Changed handling of ThingSpeak and ESP8266 so that a new module (class)
is added for handling updating of ThingSpeak. Some more cleanup needed,
but code has now run successfully in "live" tesing in greenhouse for
24h.