I2C changes and bugfixes
Fixed issue with "sensor offline" of AM2320, which was
caused because of collision with Wire configuration for
I2C between Omega and Arduino. Changed to different AM2320
library, which uses softWire and can be defined to used
different pins for Wire communication.

Changed also python test script to request for data set 2
(since only limted amount fits in one request). The test script
is now tested with real code, not the thest one...